+++ backend/substrapp/tests/views/test_views_algo.py
@@ -244,28 +244,10 @@ class AlgoViewTests(APITestCase):
             for field in ("description", "algorithm"):
                 self.assertEqual(result[field]["storage_address"], algo[field]["storage_address"])
 
-    def test_algo_list_search_filter(self):
-        """Filter algo on key."""
-        key = self.expected_results[0]["key"]
-        params = urlencode({"search": f"algo:key:{key}"})
-        response = self.client.get(f"{self.url}?{params}", **self.extra)
-        self.assertEqual(
-            response.json(), {"count": 1, "next": None, "previous": None, "results": self.expected_results[:1]}
-        )
-
     def test_algo_list_filter(self):
         """Filter algo on key."""
         key = self.expected_results[0]["key"]
         params = urlencode({"key": key})
-        response = self.client.get(f"{self.url}?{params}", **self.extra)
-        self.assertEqual(
-            response.json(), {"count": 1, "next": None, "previous": None, "results": self.expected_results[:1]}
-        )
-
-    def test_algo_list_search_filter_and(self):
-        """Filter algo on key and owner."""
-        key, owner = self.expected_results[0]["key"], self.expected_results[0]["owner"]
-        params = urlencode({"search": f"algo:key:{key},algo:owner:{owner}"})
         response = self.client.get(f"{self.url}?{params}", **self.extra)
         self.assertEqual(
             response.json(), {"count": 1, "next": None, "previous": None, "results": self.expected_results[:1]}
@@ -280,16 +262,6 @@ class AlgoViewTests(APITestCase):
             response.json(), {"count": 1, "next": None, "previous": None, "results": self.expected_results[:1]}
         )
 
-    def test_algo_list_search_filter_in(self):
-        """Filter algo in key_0, key_1."""
-        key_0 = self.expected_results[0]["key"]
-        key_1 = self.expected_results[1]["key"]
-        params = urlencode({"search": f"algo:key:{key_0},algo:key:{key_1}"})
-        response = self.client.get(f"{self.url}?{params}", **self.extra)
-        self.assertEqual(
-            response.json(), {"count": 2, "next": None, "previous": None, "results": self.expected_results[:2]}
-        )
-
     def test_algo_list_filter_in(self):
         """Filter algo in key_0, key_1."""
         key_0 = self.expected_results[0]["key"]
@@ -299,52 +271,6 @@ class AlgoViewTests(APITestCase):
         self.assertEqual(
             response.json(), {"count": 2, "next": None, "previous": None, "results": self.expected_results[:2]}
         )
-
-    def test_algo_list_search_filter_or(self):
-        """Filter algo on key_0 or key_1."""
-        key_0 = self.expected_results[0]["key"]
-        key_1 = self.expected_results[1]["key"]
-        params = urlencode({"search": f"algo:key:{key_0}-OR-algo:key:{key_1}"})
-        response = self.client.get(f"{self.url}?{params}", **self.extra)
-        self.assertEqual(
-            response.json(), {"count": 2, "next": None, "previous": None, "results": self.expected_results[:2]}
-        )
-
-    def test_algo_list_search_filter_or_and(self):
-        """Filter algo on (key_0 and owner_0) or (key_1 and owner_1)."""
-        key_0, owner_0 = self.expected_results[0]["key"], self.expected_results[0]["owner"]
-        key_1, owner_1 = self.expected_results[1]["key"], self.expected_results[1]["owner"]
-        params = urlencode(
-            {"search": f"algo:key:{key_0},algo:owner:{owner_0}-OR-algo:key:{key_1},algo:owner:{owner_1}"}
-        )
-        response = self.client.get(f"{self.url}?{params}", **self.extra)
-        self.assertEqual(
-            response.json(), {"count": 2, "next": None, "previous": None, "results": self.expected_results[:2]}
-        )
-
-    @parameterized.expand(
-        [
-            ("ALGO_SIMPLE",),
-            ("ALGO_AGGREGATE",),
-            ("ALGO_COMPOSITE",),
-            ("ALGO_METRIC",),
-            ("ALGO_XXX",),  # invalid
-        ]
-    )
-    def test_algo_list_search_filter_by_category(self, category):
-        """Filter algo on category."""
-        filtered_algos = [task for task in self.expected_results if task["category"] == category]
-        params = urlencode({"search": f"algo:category:{category}"})
-        url = self.metric_url if category == "ALGO_METRIC" else self.url
-        response = self.client.get(f"{url}?{params}", **self.extra)
-
-        if category != "ALGO_XXX":
-            self.assertEqual(
-                response.json(),
-                {"count": len(filtered_algos), "next": None, "previous": None, "results": filtered_algos},
-            )
-        else:
-            self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
 
     @parameterized.expand(
         [
@@ -389,19 +315,6 @@ class AlgoViewTests(APITestCase):
     def test_algo_match(self):
         """Match algo on part of the name."""
         params = urlencode({"match": "le al"})
-        response = self.client.get(f"{self.url}?{params}", **self.extra)
-        self.assertEqual(
-            response.json(), {"count": 1, "next": None, "previous": None, "results": self.expected_results[:1]}
-        )
-
-    def test_algo_match_and_search_filter(self):
-        """Match algo with filter."""
-        params = urlencode(
-            {
-                "search": f"algo:key:{self.expected_results[0]['key']}",
-                "match": "le al",
-            }
-        )
         response = self.client.get(f"{self.url}?{params}", **self.extra)
         self.assertEqual(
             response.json(), {"count": 1, "next": None, "previous": None, "results": self.expected_results[:1]}