+++ backend/substrapp/tests/views/test_views_datamanager.py
@@ -193,28 +193,10 @@ class DataManagerViewTests(APITestCase):
             for field in ("description", "opener"):
                 self.assertEqual(result[field]["storage_address"], data_manager[field]["storage_address"])
 
-    def test_datamanager_list_search_filter(self):
-        """Filter datamanager on key."""
-        key = self.expected_results[0]["key"]
-        params = urlencode({"search": f"dataset:key:{key}"})
-        response = self.client.get(f"{self.url}?{params}", **self.extra)
-        self.assertEqual(
-            response.json(), {"count": 1, "next": None, "previous": None, "results": self.expected_results[:1]}
-        )
-
     def test_datamanager_list_filter(self):
         """Filter datamanager on key."""
         key = self.expected_results[0]["key"]
         params = urlencode({"key": key})
-        response = self.client.get(f"{self.url}?{params}", **self.extra)
-        self.assertEqual(
-            response.json(), {"count": 1, "next": None, "previous": None, "results": self.expected_results[:1]}
-        )
-
-    def test_datamanager_list_search_filter_and(self):
-        """Filter datamanager on key and owner."""
-        key, owner = self.expected_results[0]["key"], self.expected_results[0]["owner"]
-        params = urlencode({"search": f"dataset:key:{key},dataset:owner:{owner}"})
         response = self.client.get(f"{self.url}?{params}", **self.extra)
         self.assertEqual(
             response.json(), {"count": 1, "next": None, "previous": None, "results": self.expected_results[:1]}
@@ -229,47 +211,11 @@ class DataManagerViewTests(APITestCase):
             response.json(), {"count": 1, "next": None, "previous": None, "results": self.expected_results[:1]}
         )
 
-    def test_datamanager_list_search_filter_in(self):
-        """Filter datamanager in key_0, key_1."""
-        key_0 = self.expected_results[0]["key"]
-        key_1 = self.expected_results[1]["key"]
-        params = urlencode({"search": f"dataset:key:{key_0},dataset:key:{key_1}"})
-        response = self.client.get(f"{self.url}?{params}", **self.extra)
-        self.assertEqual(
-            response.json(), {"count": 2, "next": None, "previous": None, "results": self.expected_results[:2]}
-        )
-
     def test_datamanager_list_filter_in(self):
         """Filter datamanager in key_0, key_1."""
         key_0 = self.expected_results[0]["key"]
         key_1 = self.expected_results[1]["key"]
         params = urlencode({"key": ",".join([key_0, key_1])})
-        response = self.client.get(f"{self.url}?{params}", **self.extra)
-        self.assertEqual(
-            response.json(), {"count": 2, "next": None, "previous": None, "results": self.expected_results[:2]}
-        )
-
-    def test_datamanager_list_search_filter_or(self):
-        """Filter datamanager on key_0 or key_1."""
-        key_0 = self.expected_results[0]["key"]
-        key_1 = self.expected_results[1]["key"]
-        params = urlencode({"search": f"dataset:key:{key_0}-OR-dataset:key:{key_1}"})
-        response = self.client.get(f"{self.url}?{params}", **self.extra)
-        self.assertEqual(
-            response.json(), {"count": 2, "next": None, "previous": None, "results": self.expected_results[:2]}
-        )
-
-    def test_datamanager_list_search_filter_or_and(self):
-        """Filter datamanager on (key_0 and owner_0) or (key_1 and owner_1)."""
-        key_0, owner_0 = self.expected_results[0]["key"], self.expected_results[0]["owner"]
-        key_1, owner_1 = self.expected_results[1]["key"], self.expected_results[1]["owner"]
-        params = urlencode(
-            {
-                "search": (
-                    f"dataset:key:{key_0},dataset:owner:{owner_0}" f"-OR-dataset:key:{key_1},dataset:owner:{owner_1}"
-                )
-            }
-        )
         response = self.client.get(f"{self.url}?{params}", **self.extra)
         self.assertEqual(
             response.json(), {"count": 2, "next": None, "previous": None, "results": self.expected_results[:2]}
@@ -295,19 +241,6 @@ class DataManagerViewTests(APITestCase):
     def test_datamanager_match(self):
         """Match datamanager on part of the name."""
         params = urlencode({"match": "manager fo"})
-        response = self.client.get(f"{self.url}?{params}", **self.extra)
-        self.assertEqual(
-            response.json(), {"count": 1, "next": None, "previous": None, "results": self.expected_results[:1]}
-        )
-
-    def test_datamanager_match_and_search_filter(self):
-        """Match datamanager with filter."""
-        params = urlencode(
-            {
-                "search": f"dataset:key:{self.expected_results[0]['key']}",
-                "match": "manager fo",
-            }
-        )
         response = self.client.get(f"{self.url}?{params}", **self.extra)
         self.assertEqual(
             response.json(), {"count": 1, "next": None, "previous": None, "results": self.expected_results[:1]}