+++ CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - deprecated metadata was used during event processing (#1116)
 
+### Changed
+
+- Removed the search parameter from API (#1123)
+
 ## [0.19.0] 2022-06-20
 
 ### Fixed